# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Oxter API",
        "description": "Документация",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/favorite/add": {
            "post": {
                "tags": [
                    "favorite"
                ],
                "summary": "Добавить в избранное",
                "description": "Добавить в избранное",
                "operationId": "addfavorite",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "api_token, itemid"
                                ],
                                "properties": {
                                    "api_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "R6efxg145osgPcJpb2LTXUXy1rcKezAAuPGYvdH5fFogUqT3xAGk06An6qCW"
                                    },
                                    "itemid": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CallBack"
                    }
                }
            }
        },
        "/api/favorite/view": {
            "post": {
                "tags": [
                    "favorite"
                ],
                "summary": "Посмотреть избранные товары",
                "description": "Посмотреть избранные товары",
                "operationId": "viewfavorite",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "api_token"
                                ],
                                "properties": {
                                    "api_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "R6efxg145osgPcJpb2LTXUXy1rcKezAAuPGYvdH5fFogUqT3xAGk06An6qCW"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CallBack"
                    }
                }
            }
        },
        "/api/favorite/delete": {
            "post": {
                "tags": [
                    "favorite"
                ],
                "summary": "Удалить избранный товар",
                "description": "Удалить избранный товар",
                "operationId": "deletefavorite",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "api_token, itemid"
                                ],
                                "properties": {
                                    "api_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "R6efxg145osgPcJpb2LTXUXy1rcKezAAuPGYvdH5fFogUqT3xAGk06An6qCW"
                                    },
                                    "itemid": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CallBack"
                    }
                }
            }
        },
...
    }
}